Added msub options for when email notifications occur and the address to which they should be sent. These options can be specified in the users config file.
Since these email parameters are optional in the config file but are searched for by scheduling.rb as config methods, Pipeline::Config's method_missing method was modified to catch them. Setting defaults elsewhere (such as in Config.initialize), but @config appears to be a hash there and later type SampleObject.